First step of config isolation
