+++ pkg/collector/README.md
@@ -1,7 +1,7 @@
 # package `collector`
 
 This package is responsible for providing any resource and functionality needed in a check
-life cycle: retrieving a configuration, create a check instance, schedule a check, running
+life cycle: create a check instance, schedule a check, running
 a check are all operations implemented in this package or one of the subpackages.
 
 ## Metadata
@@ -16,7 +16,6 @@ For further details, please refer to the specific READMEs:
 * [check](check/README.md)
 * [corechecks](corechecks/README.md)
 * [metadata](metadata/README.md)
-* [providers](providers/README.md)
 * [py](py/README.md)
 * [runner](runner/README.md)
 * [scheduler](scheduler/README.md)
